# Use AirflowInstance.get_task_infos to improve performance.